refactor(codegen): extract locus declaration into locus/decl.rs (Round 4b)
2 fns lifted into LocusDeclare<'ctx>: declare_locus_struct (Phase A
locus struct + field layout + capacity slot synthesis) and
declare_locus_methods (Phase A method signatures: birth, run,
dissolve, accept, modes, user fns, closure-eval synthetic fns).

The trait is named LocusDeclare (not LocusDecl) to avoid collision
with the AST node type `LocusDecl`.

Surface bumps to pub(crate):
- Cx fields: pinned_locus_types
- Cx methods: type_expr_to_codegen_ty, resolve_generic_struct_path,
  view_struct_ty
- Free fns: collect_sum_calls, count_self_field_accesses_in_locus,
  infer_accumulator_inner_type, locus_arena_elidable,
  locus_reads_self_children, param_value
- Consts: CHILDREN_CAP
- Enums: AccumulatorKind, ParamValue, SyncMode, DefaultInit
- Structs: AccumulatorSlot (+ fields), CapacitySlotLayout (residual
  fields: indexed_by, ring_buffer_cap, sync_mode)
- TypeInfo::field_order, TypeInfo::defaults

codegen.rs: 33895 -> 32205 (-1690 lines).
