+++ crates/hale-codegen/src/locus/mod.rs
@@ -1,4 +1,5 @@
 //! Locus codegen (lifecycle, methods, dissolve, arenas, placement,
 //! return path). Round 4 of the codegen model-org refactor.
 
+pub(crate) mod decl;
 pub(crate) mod dissolve;
